Adjusts styling of exchange screens and renames folder
