+++ src/frontend/screens/Exchange/NoWifiDisplay.tsx
